+++ src/frontend/components/mod.rs
@@ -1,0 +1,6 @@
+mod conversation;
+mod message;
+mod sidebar;
+
+pub(crate) use conversation::Conversation;
+pub(crate) use sidebar::Sidebar;
