+++ src/frontend/mod.rs
@@ -1,0 +1,2 @@
+pub mod components;
+pub mod views;
